requirements.txt
# ── Core ML ──────────────────────────────────────
lightgbm>=4.3
xgboost>=2.0
scikit-learn>=1.4
imbalanced-learn>=0.12
pandas>=2.2
numpy>=1.26
scipy>=1.13

# ── Data / IO ─────────────────────────────────────
pyarrow>=15.0          # parquet support
requests>=2.31
kaggle>=1.6            # optional: Kaggle API for dataset download

# ── API backend ───────────────────────────────────
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.7

# ── Visualisation / Notebooks ─────────────────────
matplotlib>=3.8
seaborn>=0.13
shap>=0.45
jupyter>=1.0
nbconvert>=7.0

# ── Utilities ─────────────────────────────────────
pyyaml>=6.0
joblib>=1.4            # model serialisation
python-dotenv>=1.0

# ── Testing ───────────────────────────────────────
pytest>=8.2
httpx>=0.27            # FastAPI test client

# ── Optional monitoring ───────────────────────────
# prometheus-client>=0.20   # uncomment if emitting .prom metrics
